AML-3882: migrate WhenIGetAccountUsersWithKnownIds; except for the ExpectedControllerType assertion.
